Fixed some warnings on Windows.

+++ cppparser.cpp
@@ -153,7 +153,7 @@ CppParser::ByteArray CppParser::readFile(const char* filename)
   if (in)
   {
     in.seekg(0, std::ios::end);
-    size_t size = in.tellg();
+    size_t size = static_cast<size_t>(in.tellg());
     contents.resize(size + 3); // For adding last 2 nulls and a new line.
     in.seekg(0, std::ios::beg);
     in.read(&contents[0], size);
